yolo-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
ultralytics==8.3.0
httpx==0.27.2
Pillow==10.4.0
python-multipart==0.0.12
